Move WorkflowConnection to Connection namespace
