+++ src/Processor/ExclusiveGatewayProcessor.php
@@ -2,9 +2,9 @@
 
 namespace Phlow\Processor;
 
+use Phlow\Connection\WorkflowConnection;
 use Phlow\Engine\Exchange;
 use Phlow\Engine\ExpressionEngine;
-use Phlow\Model\WorkflowConnection;
 use Phlow\Model\WorkflowNode;
 
 /**
